Indirect left recursion elimination - t_symbol
To obtain a LL(1) grammar, one must eliminate left recursion, including
indirect one.
Implement the following transformation :
'A' being a symbol preceding 'B', with these productions :
A => C D | E F | Z ;
B being with these production :
B => X H | A T | A J | R A ;
Should become
B => X H | R A
    | C D T | E F T | Z T
    | C D J | E F J | Z J
    ;
That is, every production in B where A is the first symbol should be
expanded into a list a productions using each A-production with the
remainder of the B production to produces a new one.